Rename adapter module to adapters
