+++ tardis/adapters/sites/openstack.py
